deprecated reducer almost Angular free
