+++ src/map/ducks/search-results/map-search-results.js
@@ -57,6 +57,3 @@ export const getMapSearchResults = (location, user) => ({
   location,
   user
 });
-
-window.reducers = window.reducers || {};
-window.reducers.MapSearchResultsReducer = MapSearchResultsReducer;
