+++ src/reducers/deprecated/deprecated-reducer.js
@@ -1,4 +1,4 @@
-import DetailReducer from '../details';
+import DetailsReducer from '../details';
 
 import DataSelectionReducer from '../data-selection-reducers';
 import MapPreviewPanelReducer from '../../map/ducks/preview-panel/map-preview-panel';
@@ -6,6 +6,9 @@ import PageReducer from '../page-reducers';
 import homeReducer from '../home-reducers';
 import filtersReducers from './filters-reducers';
 import straatbeeldReducers from './straatbeeld-reducers';
+import MapSearchResultsReducer from '../../map/ducks/search-results/map-search-results';
+import MapClickLocationReducer from '../../map/ducks/click-location/map-click-location';
+import searchReducers from './search-reducers';
 
 export default (oldState, action) => {
   /**
@@ -18,56 +21,56 @@ export default (oldState, action) => {
    *
    */
   const UrlReducers = window.reducers.UrlReducers;
-  // const MapPreviewPanelReducer = $window.reducers.MapPreviewPanelReducer;
-  // const MapSearchResultsReducer = $window.reducers.MapSearchResultsReducer;
-  // const MapClickLocationReducer = $window.reducers.MapClickLocationReducer;
-  //
-  // const detailReducers = {
-  //   FETCH_DETAIL: DetailsReducer,
-  //   SHOW_DETAIL: DetailsReducer,
-  //   DETAIL_FULLSCREEN: DetailsReducer
-  // };
-  //
-  // const mapSearchResultsReducers = {
-  //   FETCH_MAP_SEARCH_RESULTS_FAILURE: MapSearchResultsReducer,
-  //   FETCH_MAP_SEARCH_RESULTS_REQUEST: MapSearchResultsReducer,
-  //   FETCH_MAP_SEARCH_RESULTS_SUCCESS: MapSearchResultsReducer
-  // };
-  //
-  // const mapClickLocationReducers = {
-  //   SET_MAP_CLICK_LOCATION: MapClickLocationReducer
-  // };
-  //
+  // const MapPreviewPanelReducer = $window.reducers.MapPreviewPanelReducer;      // DONE
+  // const MapSearchResultsReducer = $window.reducers.MapSearchResultsReducer;    // DONE
+  // const MapClickLocationReducer = $window.reducers.MapClickLocationReducer;    // DONE
+
+  const detailReducers = { // TODO: try moving to root reducer
+    FETCH_DETAIL: DetailsReducer,
+    SHOW_DETAIL: DetailsReducer,
+    DETAIL_FULLSCREEN: DetailsReducer
+  };
+
+  const mapSearchResultsReducers = { // TODO: try moving to root reducer
+    FETCH_MAP_SEARCH_RESULTS_FAILURE: MapSearchResultsReducer,
+    FETCH_MAP_SEARCH_RESULTS_REQUEST: MapSearchResultsReducer,
+    FETCH_MAP_SEARCH_RESULTS_SUCCESS: MapSearchResultsReducer
+  };
+
+  const mapClickLocationReducers = { // TODO: try moving to root reducer
+    SET_MAP_CLICK_LOCATION: MapClickLocationReducer
+  };
+
   const mapPreviewPanelReducers = {
     OPEN_MAP_PREVIEW_PANEL: MapPreviewPanelReducer,
     CLOSE_MAP_PREVIEW_PANEL: MapPreviewPanelReducer,
     MAXIMIZE_MAP_PREVIEW_PANEL: MapPreviewPanelReducer
   };
 
-  // TODO: Redux: replace
-  // Warning: angular.merge is deprecated
-  // -- https://docs.angularjs.org/api/ng/function/angular.merge
   // var actions = angular.merge(
-  //   dataSelectionReducers,   // Done
-  //   detailReducers,          // Done
-  //   filtersReducers,         // Done
-  //   homeReducers,            // Done
-  //   mapPreviewPanelReducers, // Done
-  //   mapSearchResultsReducers,
-  //   mapClickLocationReducers,
-  //   pageReducers,            // Done
-  //   searchReducers,
-  //   straatbeeldReducers,     // Done
-  //   urlReducers,             // Done
+  //   dataSelectionReducers,   // DONE
+  //   detailReducers,          // DONE
+  //   filtersReducers,         // DONE
+  //   homeReducers,            // DONE
+  //   mapPreviewPanelReducers, // DONE
+  //   mapSearchResultsReducers,// DONE
+  //   mapClickLocationReducers,// DONE
+  //   pageReducers,            // DONE
+  //   searchReducers,          // DONE
+  //   straatbeeldReducers,     // DONE
+  //   urlReducers,             // DONE
   //   environment
   // );
   const actions = {
     ...DataSelectionReducer,
-    ...DetailReducer,
+    ...detailReducers,
     ...filtersReducers,
     ...homeReducer,
     ...mapPreviewPanelReducers,
+    ...mapSearchResultsReducers,
+    ...mapClickLocationReducers,
     ...PageReducer,
+    ...searchReducers,
     ...straatbeeldReducers,
     ...UrlReducers,
   };
