+++ src/reducers/deprecated/search-reducers.js
@@ -1,0 +1,114 @@
+import ACTIONS from '../../shared/actions';
+import BaseCoder from '../../../modules/shared/services/base-coder/base-coder';
+
+/**
+ * @param {Object} state
+ * @param {String} payload - A search query
+ *
+ * @returns {Object} newState
+ */
+function fetchSearchResultsByQueryReducer(state, payload) {
+  return {
+    ...state,
+    search: {
+      isLoading: true,
+      isFullscreen: true,
+      query: payload || null,
+      location: null,
+      category: null,
+      numberOfResults: null
+    },
+    ui: typeof state.ui === 'object' ? {
+      ...state.ui,
+      isMapPanelVisible: false,
+      isMapFullscreen: false
+    } : state.ui,
+    page: typeof state.page === 'object' ? {
+      ...state.page,
+      name: null,
+      type: null
+    } : state.page,
+    detail: null,
+    straatbeeld: null,
+    dataSelection: null
+  };
+}
+
+/**
+ * @param {Object} state
+ * @param {Array} payload - A location, e.g. [52.123, 4.789]
+ *
+ * @returns {Object} newState
+ */
+function fetchSearchResultsByLocationReducer(state, payload) {
+  return {
+    ...state,
+    search: {
+      isLoading: true,
+      isFullscreen: false,
+      query: null,
+      location: BaseCoder.toPrecision(payload, 7),
+      category: null,
+      numberOfResults: null
+    },
+    map: typeof state.map === 'object' ? {
+      ...state.map,
+      geometry: []
+    } : state.map,
+    page: typeof state.page === 'object' ? {
+      ...state.page,
+      name: null
+    } : state.page,
+    detail: null,
+    straatbeeld: null,
+    dataSelection: null
+  };
+}
+
+/**
+ * @param {Object} state
+ * @param {String} payload - A reference to the slug of a SEARCH_CONFIG endpoint
+ *
+ * @returns {Object} newState
+ */
+function fetchSearchResultsCategoryReducer(state, payload) {
+  return {
+    ...state,
+    search: typeof state.search === 'object' ? {
+      ...state.search,
+      isLoading: true,
+      category: payload,
+      numberOfResults: null
+    } : state.search
+  };
+}
+
+/**
+ * @param {Object} oldState
+ * @param {String} payload - The number of search results
+ *
+ * @returns {Object} newState
+ */
+function showSearchResultsReducer(state, payload) {
+  return {
+    ...state,
+    search: typeof state.search === 'object' ? {
+      ...state.search,
+      isLoading: false,
+      numberOfResults: payload
+    } : state.search,
+    map: typeof state.map === 'object' ? {
+      ...state.map,
+      isLoading: false
+    } : state.map
+  };
+}
+
+var reducers = {};
+
+reducers[ACTIONS.FETCH_SEARCH_RESULTS_BY_QUERY.id] = fetchSearchResultsByQueryReducer;
+reducers[ACTIONS.FETCH_SEARCH_RESULTS_BY_LOCATION.id] = fetchSearchResultsByLocationReducer;
+reducers[ACTIONS.FETCH_SEARCH_RESULTS_CATEGORY.id] = fetchSearchResultsCategoryReducer;
+reducers[ACTIONS.SHOW_SEARCH_RESULTS.id] = showSearchResultsReducer;
+
+export default reducers;
